requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==1.20.0
google-cloud-storage